Update mid layer crash recovery

+++ src/audioCore/Crash.cpp
@@ -9,14 +9,14 @@ extern "C" void audioCoreFreeProjectDataBlock(char* mem) {
 	auto data = std::unique_ptr<google::protobuf::Message>((google::protobuf::Message*)mem);
 }
 
-extern "C" int audioCoreGetProjectDataSize(const char* mem) {
+extern "C" size_t audioCoreGetProjectDataSize(const char* mem) {
 	if (auto data = (google::protobuf::Message*)mem) {
 		return data->ByteSizeLong();
 	}
 	return 0;
 }
 
-extern "C" int audioCoreCopyProjectData(char* dst, int size, const char* mem) {
+extern "C" int audioCoreCopyProjectData(char* dst, size_t size, const char* mem) {
 	if (auto data = (google::protobuf::Message*)mem) {
 		return data->SerializeToArray(dst, size);
 	}
